Extract helper functions to halpers directory
Related to #44
